fix(weaponbutton) : 무기 레벨업 버튼 에러 해결
무기 3개의 레벨업 구현

+++ Assets/scripts/message.cs
@@ -4,7 +4,7 @@ namespace AJH{
     public class message : MonoBehaviour
     {
         [Header("데미지")]
-        public int damage = 5;
+        public float damage = 5;
 
 
         private Collider2D _myCollider;
